comment: generateBasicToken 함수 JSDoc 주석 추가
