--- src/utils/auth.ts
@@ -1,7 +1,0 @@
-export function generateBasicToken(
-  userId: string,
-  userPassword: string
-): string {
-  const token = btoa(`${userId}:${userPassword}`);
-  return `Basic ${token}`;
-}
